Remove static config and integrate with CI
