+++ GitVersion.yml
@@ -1,0 +1,13 @@
+mode: ContinuousDelivery
+ignore:
+  sha: []
+branches:
+  develop:
+    increment: Patch
+  release:
+    regex: release[/-]
+    mode: ContinuousDelivery
+    tag: rc
+    increment: Patch
+    prevent-increment-of-merged-branch-version: true
+
