+++ appveyor.yml
@@ -1,0 +1,11 @@
+# common configuration for ALL branches
+image: Visual Studio 2017
+test: off
+cache:
+- tools -> build.cake
+
+build_script:
+  - ps: .\build.ps1 -Target "Deploy"
+
+artifacts:
+  - path: 'artifacts\*.nupkg'
