+++ iDealAdvanced.sln
@@ -7,9 +7,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		AssemblyInfo.Global.cs = AssemblyInfo.Global.cs
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "iDealAdvancedConnector", "iDealAdvancedConnector\iDealAdvancedConnector.csproj", "{31C95E7E-5796-4EBF-A0F0-346850E6FA78}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "iDealAdvancedConnector", "src\iDealAdvancedConnector\iDealAdvancedConnector.csproj", "{31C95E7E-5796-4EBF-A0F0-346850E6FA78}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "iDealSampleConsole", "iDealSampleConsole\iDealSampleConsole.csproj", "{20DB033E-0DAD-4285-B7EA-40DA11CF7DEB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "iDealSampleConsole", "src\iDealSampleConsole\iDealSampleConsole.csproj", "{20DB033E-0DAD-4285-B7EA-40DA11CF7DEB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
