--- iDealAdvancedConnector/iDealAdvancedConnector.csproj
@@ -1,12 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="System.Security.Cryptography.Xml" Version="4.5.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Messages\itt-acq.xsd" />
-    <EmbeddedResource Include="Messages\xmldsigcore-schema.xsd" />
-</ItemGroup>
-</Project>
