+++ src/iDealAdvancedConnector/Constants/Constants.cs
