+++ src/iDealAdvancedConnector/Data/IDealException.cs
