+++ src/iDealAdvancedConnector/Data/Issuer.cs
