+++ src/iDealAdvancedConnector/Data/Transaction.cs
