+++ src/iDealAdvancedConnector/Messages/AcquirerStatusReqMerchant.cs
