+++ src/iDealAdvancedConnector/Messages/Constants.cs
