+++ src/iDealAdvancedConnector/Messages/DirectoryReq.cs
