+++ src/iDealAdvancedConnector/Messages/DirectoryReqMerchant.cs
