+++ src/iDealAdvancedConnector/Messages/IMerchant.cs
