+++ src/iDealAdvancedConnector/Messages/IRequest.cs
