+++ src/iDealAdvancedConnector/Messages/itt-acq.cs
