+++ src/iDealAdvancedConnector/Security/AcceptSSLCertificate.cs
