+++ src/iDealAdvancedConnector/Security/XsdValidation.cs
