+++ src/iDealAdvancedConnector/SerializationHelper.cs
