+++ src/iDealAdvancedConnector/UTF8StringWriter.cs
