+++ src/iDealAdvancedConnector/XmlSignature/XmlSignature.cs
