+++ src/iDealSampleConsole/appsettings.json
@@ -1,0 +1,15 @@
+{
+  "IDealConnector": {
+    "AcquirerCertificate": "Internet Widgits Pty Ltd",
+    "AcquirerTimeout": "10",
+    "ExpirationPeriod": "PT10M",
+    "MerchantId": "399965650",
+    "MerchantReturnURL": "http://localhost/iDealSample/PageRequestTransactionStatus.aspx",
+    "ClientCertificate": "Maxcode",
+    "SubId": "0",
+    "AcquirerURL": "",
+    "AcquirerDirectoryURL": "",
+    "AcquirerTransactionURL": "",
+    "AcquirerTransactionStatusURL": ""
+  }
+}
